Introduce BaseContentItem for shared fields
