+++ app/Data/NewsItem.php
@@ -2,22 +2,23 @@
 
 namespace App\Data;
 
-use Illuminate\Support\Str;
 use App\Services\DrupalApi;
 
-class NewsItem
+class NewsItem extends BaseContentItem
 {
     public function __construct(
         public string $id,
-        public string $title,
-        public ?string $titleEn,
-        public ?string $bodyHtml,
-        public ?string $bodyHtmlEn,
+        string $title,
+        ?string $titleEn,
+        ?string $bodyHtml,
+        ?string $bodyHtmlEn,
         public ?string $date,
-        public string $lang,
+        string $lang,
         public bool $showDate = false,
         public bool $showTime = false,
-    ) {}
+    ) {
+        parent::__construct($title, $titleEn, $bodyHtml, $bodyHtmlEn, $lang);
+    }
     
     public static function fromDrupal(array $item): self
     {
@@ -37,18 +38,4 @@ class NewsItem
         );
     }      
 
-    public function slug(): string
-    {
-        return Str::slug($this->title);
-    }
-
-    public function localizedTitle(string $locale): string
-    {
-        return $locale === 'en' && $this->titleEn ? $this->titleEn : $this->title;
-    }
-
-    public function localizedBody(string $locale): ?string
-    {
-        return $locale === 'en' && $this->bodyHtmlEn ? $this->bodyHtmlEn : $this->bodyHtml;
-    }
 }
